Comment updates and refactoring
