+++ Source/DafnyLanguageServer.Test/Lookup/HoverVerificationTest.cs
@@ -31,7 +31,7 @@ namespace Microsoft.Dafny.LanguageServer.IntegrationTest.Lookup {
       notificationReceiver = new();
       client = await InitializeClient(options => {
         options
-          .AddHandler(DafnyRequestNames.CompilationStatus, NotificationHandler.For<CompilationStatusParams>(notificationReceiver.NotificationReceived));
+          .AddHandler(NotificationHandler.For<CompilationStatusParams>(notificationReceiver.NotificationReceived));
       });
     }
