+++ Source/DafnyLanguageServer/Workspace/Notifications/CompilationStatusParams.cs
@@ -4,10 +4,12 @@ using OmniSharp.Extensions.LanguageServer.Protocol;
 
 namespace Microsoft.Dafny.LanguageServer.Workspace.Notifications {
   /// <summary>
-  /// DTO used to communicate the current compilation status to the LSP client.
+  /// Obsolete, replaced by FileVerificationStatus
+  ///
+  /// Describes the current compilation status. Useful for a status bar.
   /// </summary>
   [Method(DafnyRequestNames.CompilationStatus, Direction.ServerToClient)]
-  public class CompilationStatusParams : IRequest, IRequest<Unit> {
+  public class CompilationStatusParams : IRequest {
     /// <summary>
     /// Gets the URI of the document whose verification completed.
     /// </summary>
